+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [3.0.0] - 2021-08-10
+### Added
+- add requirement for number of signatures
+
+### Changed
+- merge `StakingBank` and `ValidatorRegistry`
+
 ## [2.0.7] - 2021-07-13
 ### Fixed
 - fix code verification contract